# openapi.yaml
openapi: 3.0.3
info:
  title: NomosX API
  description: Agentic think tank API — research intelligence platform
  version: 2.0.0
  contact:
    name: NomosX Team
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://nomosx.netlify.app/api
    description: Production
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: nomosx-session
      description: JWT session cookie (set by /auth/login)
    AdminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin key for privileged endpoints
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Brief:
      type: object
      properties:
        id: { type: string }
        question: { type: string }
        html: { type: string }
        createdAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              retryAfter: { type: integer }
paths:
...
  /briefs:
    get:
      summary: List briefs (auth required)
      tags: [Briefs]
      security: [{ SessionCookie: [] }]
      responses:
        '200': { description: List of briefs }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Generate a brief (auth + rate limited 5/min)
      tags: [Briefs]
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
      responses:
        '200':
          description: Brief generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Brief' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ============ ANALYSIS ============
...
  /newsletter/subscribe:
    post:
      summary: Subscribe to newsletter
      tags: [Newsletter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                source: { type: string }
      responses:
        '200': { description: Subscribed }

  # ============ DIGESTS ============
...
  /cron/weekly-briefs:
    post:
      summary: Send weekly brief digest
      tags: [Cron]
      security: [{ AdminKey: [] }]
      responses:
        '200': { description: Weekly briefs sent }